Initial numberline refactor.
