+++ src/plugins/numberline/components/numberline-tile.scss
@@ -1,5 +1,5 @@
 
-@import "../../components/vars.sass";
+@import "../../../components/vars.sass";
 
 .numberline-tool {
   width: 100%;
