+++ src/plugins/numberline/components/numberline-toolbar.scss
@@ -1,4 +1,4 @@
-@import "../../components/vars.sass";
+@import "../../../components/vars.sass";
 
 .numberline-toolbar{
   position: absolute;
