+++ src/plugins/numberline/components/numberline-toolbar.tsx
@@ -1,10 +1,11 @@
 import { observer } from "mobx-react";
 import React from "react";
 import ReactDOM from "react-dom";
-import { IFloatingToolbarProps, useFloatingToolbarLocation
-        } from "../../components/tiles/hooks/use-floating-toolbar-location";
-import { useSettingFromStores } from "../../hooks/use-stores";
+
 import { PlacePointButton, ClearPointsButton, DeletePointButton } from "./numberline-toolbar-buttons";
+import { IFloatingToolbarProps, useFloatingToolbarLocation
+        } from "../../../components/tiles/hooks/use-floating-toolbar-location";
+import { useSettingFromStores } from "../../../hooks/use-stores";
 
 import "./numberline-toolbar.scss";
 
@@ -41,7 +42,6 @@ export const NumberlineToolbar: React.FC<INumberlineToolbarProps> = observer((pr
         return <DeletePointButton key={toolName} onClick={handleDeletePoint} />;
     }
   };
-
 
   return documentContent
     ? ReactDOM.createPortal(
